Use react-resolver src instead of dist
