+++ examples/react-v0.13/src/server.js
@@ -1,54 +1,7 @@
-import express from "express";
-import React from "react";
-import { Resolver } from "react-resolver";
-import Router from "react-router";
-import ServerLocation from "react-router-server-location";
+require("babel/register")({ only: /src|react-resolver/ });
 
-import routes from "./routes";
+var app = require("./app");
 
-express()
-  // Let React handle all routes
-  .get("*", function(req, res) {
-    const location = new ServerLocation(req, res);
-
-    Router.create({ location, routes }).run(function(Handler, state) {
-      // Redirect any unknown to home
-      if (!state.routes.length) {
-        return res.redirect("/");
-      }
-
-      const notFound = state.routes.filter(({ name }) => name === "404").length;
-
-      Resolver
-        .resolve(() => <Handler {...state} />)
-        .then(({ Resolved, data }) => {
-          res
-            .status(notFound ? 404 : 200)
-            .send(`
-              <!DOCTYPE html>
-              <html lang="en">
-              <head>
-                <meta charset="UTF-8">
-                <title>Stargazers Demo – React Resolver</title>
-                <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.0/css/materialize.min.css">
-              </head>
-              <body>
-                <div id="app">${React.renderToString(<Resolved />)}</div>
-
-                <script src="/client.min.js" async defer></script>
-                <script>window.__REACT_RESOLVER_PAYLOAD__ = ${JSON.stringify(data)}</script>
-              </body>
-              </html>
-            `)
-          ;
-        })
-        .catch((error) => {
-          res.status(500).send(error);
-        })
-      ;
-    });
-  })
-  .listen(3000, function() {
-    console.info("✅  Node server started at http://localhost:3000/");
-  })
-;
+app.listen(3000, function() {
+  console.info("✅  Node server started at http://localhost:3000/");
+});
