+++ examples/react-v0.13/watch.js
@@ -5,6 +5,22 @@ var piping = require("piping")({ hook: true });
 var webpack = require("webpack");
 var WebpackDevServer = require("webpack-dev-server");
 
+config.debug = true;
+config.devtool = "#eval-source-map";
+
+config.entry.client.unshift(
+  "webpack-dev-server/client?http://localhost:8080",
+  "webpack/hot/only-dev-server"
+);
+
+config.plugins.unshift(
+  // Wrap builds with HMR
+  new webpack.HotModuleReplacementPlugin(),
+
+  // Prevent errors from causing a reload
+  new webpack.NoErrorsPlugin()
+);
+
 if (piping) {
   require("./src/server");
 } else {
